requirements.txt
asgiref==3.5.0
Django==4.0.3
django-cors-headers==3.11.0
django-dotenv==1.4.2
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
PyJWT==2.3.0
pytz==2021.3
sqlparse==0.4.2
tzdata==2021.5
psycopg2==2.9.9
drf-yasg==1.21.7
dj-database-url==2.1.0
shortuuid==1.0.11
Pillow
setuptools
daphne
channels
channels-redis
spacy
rapidfuzz
gunicorn
whitenoise
psycopg2-binary
python-decouple